Create array-type H5 file

Function to populate hdf5 file with array type data.

write_array(
  array,
  handle,
  data_product,
  component,
  description,
  dimension_names,
  dimension_values,
  dimension_units,
  units
)

Arguments

array
an array containing the data

handle
fdp object

data_product
a string specifying the name of the data product

component
a string specifying a location within the hdf5 file

description
a string describing the data product component

dimension_names
a list where each element is a vector containing the labels associated with a particular dimension (e.g. element 1 corresponds to dimension 1, which corresponds to row names) and the name of each element describes the contents of each dimension (e.g. age classes).

dimension_values
(optional) a list of values corresponding to each dimension (e.g. list element 2 corresponds to columns)

dimension_units
(optional) a list of units corresponding to each dimension (e.g. list element 2 corresponds to columns)

units
(optional) a string specifying the units of the data as a whole

See also